requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-CORS>=4.0.0
PyPDF2>=3.0.1
reportlab>=4.2.0
Pillow>=11.1.0
stripe>=11.0.0
python-dotenv>=1.0.0
email-validator>=2.1.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
